fix some initialization bugs

+++ Arduino/PyramidOfPossibilities/Constants.h
@@ -56,8 +56,8 @@ struct RNConstants {
     const uint16_t pyramidRadius = 7209;
     const uint16_t pyramidRadiusFromGround = 7796;
 
-    const uint16_t serial2BaudRate = 38400;
-    const uint16_t usbSerialBaudRate = 115200;
+    const uint32_t serial2BaudRate = 38400;
+    const uint32_t usbSerialBaudRate = 115200;
     const uint16_t watchdogTimeout = 10000;
     
 };
